Cleanup nav utils
