+++ app/src/main/java/com/bottlerocketstudios/brarchitecture/ui/ComposeActivity.kt
@@ -9,24 +9,10 @@ import androidx.compose.material.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.lifecycle.asFlow
-import androidx.navigation.NavController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.bottlerocketstudios.compose.resources.ArchitectureDemoTheme
 import org.koin.androidx.viewmodel.ext.android.viewModel
-
-// TODO - move to new file
-@Suppress("FunctionName")
-object Routes {
-    const val Main = "main"
-    const val Home = "home"
-    const val Splash = "splash"
-    const val AuthCode = "authcode"
-    const val DevOptions = "devoptions"
-
-    // Example path with arguments; will remove after first real path with arguments is in place.
-    fun UserProfile(id: String) = "profile/{${id}}"
-}
 
 
 class ComposeActivity : ComponentActivity() {
